# sdk/go/api/openapi.yaml
openapi: 3.1.0
info:
  description: "C2PA-compliant content signing and verification infrastructure for\
    \ publishers, legal/finance firms, AI labs, and enterprises"
  title: Encypher Enterprise API
  version: 1.0.0-preview
servers:
- description: Production
  url: https://api.encypherai.com
- description: Local development
  url: http://localhost:8007
paths:
...
  /api/v1/byok/public-keys:
    get:
      description: List all registered public keys for the organization.
      operationId: list_public_keys_api_v1_byok_public_keys_get
      parameters:
      - description: Include revoked keys
        explode: true
        in: query
        name: include_revoked
        required: false
        schema:
          default: false
          description: Include revoked keys
          title: Include Revoked
          type: boolean
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicKeyListResponse"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: List public keys
      tags:
      - BYOK
      - BYOK
    post:
      description: Register a BYOK public key for signature verification.
      operationId: register_public_key_api_v1_byok_public_keys_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublicKeyRegisterRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicKeyRegisterResponse"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Register a public key
      tags:
      - BYOK
      - BYOK
  /api/v1/byok/public-keys/{key_id}:
    delete:
      description: Revoke a registered public key. Revoked keys cannot be used for
        verification.
      operationId: revoke_public_key_api_v1_byok_public_keys__key_id__delete
      parameters:
      - explode: false
        in: path
        name: key_id
        required: true
        schema:
          title: Key Id
          type: string
        style: simple
      - description: Reason for revocation
        explode: true
        in: query
        name: reason
        required: false
        schema:
          nullable: true
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: {}
                title: Response Revoke Public Key Api V1 Byok Public Keys  Key Id  Delete
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Revoke a public key
      tags:
      - BYOK
      - BYOK
...
components:
  schemas:
...
    PublicKeyListResponse:
      description: Response for listing organization's public keys.
      example:
        data:
          keys: []
          total: 0
        success: true
      properties:
        success:
          default: true
          title: Success
          type: boolean
        data:
          additionalProperties: {}
          title: Data
      title: PublicKeyListResponse
...
    HTTPValidationError:
      example:
        detail:
        - msg: msg
          loc:
          - ValidationError_loc_inner
          - ValidationError_loc_inner
          type: type
        - msg: msg
          loc:
          - ValidationError_loc_inner
          - ValidationError_loc_inner
          type: type
      properties:
        detail:
          items:
            $ref: "#/components/schemas/ValidationError"
          type: array
      title: HTTPValidationError
...
    ValidationError:
      example:
        msg: msg
        loc:
        - ValidationError_loc_inner
        - ValidationError_loc_inner
        type: type
      properties:
        loc:
          items:
            $ref: "#/components/schemas/ValidationError_loc_inner"
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    PublicKeyRegisterResponse:
      description: Response for public key registration.
      example:
        data:
          created_at: 2024-12-23T22:00:00Z
          id: pk_abc123
          is_active: true
          key_algorithm: Ed25519
          key_fingerprint: SHA256:abc123...
          key_name: Production Signing Key
          organization_id: org_xyz789
          public_key_pem: |-
            -----BEGIN PUBLIC KEY-----
            ...
            -----END PUBLIC KEY-----
        success: true
      properties:
        success:
          default: true
          title: Success
          type: boolean
        data:
          $ref: "#/components/schemas/PublicKeyInfo"
        error:
          nullable: true
          type: string
      title: PublicKeyRegisterResponse
    PublicKeyInfo:
      description: Information about a registered public key.
      properties:
        id:
          description: Key ID
          title: Id
          type: string
        organization_id:
          description: Organization that owns this key
          title: Organization Id
          type: string
        key_name:
          nullable: true
          type: string
        key_algorithm:
          description: Key algorithm
          title: Key Algorithm
          type: string
        key_fingerprint:
          description: SHA-256 fingerprint of the public key
          title: Key Fingerprint
          type: string
        public_key_pem:
          description: PEM-encoded public key
          title: Public Key Pem
          type: string
        is_active:
          default: true
          description: Whether key is active for verification
          title: Is Active
          type: boolean
        created_at:
          description: When key was registered
          format: date-time
          title: Created At
          type: string
        last_used_at:
          format: date-time
          nullable: true
          type: string
      required:
      - created_at
      - id
      - key_algorithm
      - key_fingerprint
      - organization_id
      - public_key_pem
      title: PublicKeyInfo
...
    PublicKeyRegisterRequest:
      description: Request to register a BYOK public key.
      example:
        key_algorithm: Ed25519
        key_name: Production Signing Key
        public_key_pem: |-
          -----BEGIN PUBLIC KEY-----
          MCowBQYDK2VwAyEA...
          -----END PUBLIC KEY-----
      properties:
        public_key_pem:
          description: PEM-encoded public key (Ed25519 or RSA)
          minLength: 50
          title: Public Key Pem
          type: string
        key_name:
          nullable: true
          type: string
        key_algorithm:
          default: Ed25519
          description: "Key algorithm (Ed25519, RSA-2048, RSA-4096)"
          title: Key Algorithm
          type: string
      required:
      - public_key_pem
      title: PublicKeyRegisterRequest
...
    ValidationError_loc_inner:
      anyOf:
      - type: string
      - type: integer
  securitySchemes:
    HTTPBearer:
      scheme: bearer
      type: http